refactor(coral): Rename topics feature.
