+++ coral/src/app/features/browse-topics/BrowseTopics.tsx
@@ -1,14 +1,14 @@
-import { useGetTopics } from "src/app/features/topics/hooks/list/useGetTopics";
+import { useGetTopics } from "src/app/features/browse-topics/hooks/list/useGetTopics";
 import { Pagination } from "src/app/components/Pagination";
-import SelectTeam from "src/app/features/topics/components/select-team/SelectTeam";
-import TopicList from "src/app/features/topics/components/list/TopicList";
-import { useGetTeams } from "src/app/features/topics/hooks/teams/useGetTeams";
+import SelectTeam from "src/app/features/browse-topics/components/select-team/SelectTeam";
+import TopicList from "src/app/features/browse-topics/components/list/TopicList";
+import { useGetTeams } from "src/app/features/browse-topics/hooks/teams/useGetTeams";
 import { useState } from "react";
 import { Flexbox, FlexboxItem } from "@aivenio/design-system";
 import { useSearchParams } from "react-router-dom";
 import { Environment } from "src/domain/environment";
-import SelectEnvironment from "src/app/features/topics/components/select-environment/SelectEnvironment";
-import { useGetEnvironments } from "src/app/features/topics/hooks/environment/useGetEnvironments";
+import SelectEnvironment from "src/app/features/browse-topics/components/select-environment/SelectEnvironment";
+import { useGetEnvironments } from "src/app/features/browse-topics/hooks/environment/useGetEnvironments";
 
 // Use a UUID value to represent empty option value.
 const ALL_TEAMS_VALUE = "f5ed03b4-c0da-4b18-a534-c7e9a13d1342";
