+++ coral/src/app/features/browse-topics/components/list/TopicCard.tsx
@@ -7,9 +7,9 @@ import {
   Flexbox,
   ExternalLinkButton,
 } from "@aivenio/design-system";
-import classes from "src/app/features/topics/components/list/TopicCard.module.css";
+import classes from "src/app/features/browse-topics/components/list/TopicCard.module.css";
 import { Topic } from "src/domain/topic";
-import { createTopicOverviewLink } from "src/app/features/topics/utils/create-topic-overview-link";
+import { createTopicOverviewLink } from "src/app/features/browse-topics/utils/create-topic-overview-link";
 
 type TopicCardPros = Pick<
   Topic,
