+++ coral/src/app/features/browse-topics/components/list/TopicList.test.tsx
@@ -1,7 +1,7 @@
 import { cleanup, screen, within } from "@testing-library/react";
 import { renderWithQueryClient } from "src/services/test-utils";
 import { createMockTopic } from "src/domain/topic/topic-test-helper";
-import TopicList from "src/app/features/topics/components/list/TopicList";
+import TopicList from "src/app/features/browse-topics/components/list/TopicList";
 
 const mockedTopicNames = ["Name one", "Name two", "Name three", "Name four"];
 const mockedTopics = mockedTopicNames.map((name, index) =>
